LayoutTests/fast/dom/element-attribute-js-null-expected.txt

This test setting various attributes of a elements to JavaScript null.

TEST SUCCEEDED: The value was null. [tested Node.nodeValue]
TEST SUCCEEDED: The value was null. [tested Node.prefix]
TEST SUCCEEDED: The value was the empty string. [tested Node.textContent]

TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.title]
TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.id]
TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.lang]
TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.dir]
TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.className]
TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.innerHTML]
TEST SUCCEEDED: The value was the empty string. [tested HTMLElement.innerText]
TEST SUCCEEDED: Exception (Error: SYNTAX_ERR: DOM Exception 12) was thrown as expected. [tested HTMLElement.contentEditable]

TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.charset]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.coords]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.href]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.hreflang]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.rel]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.rev]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.shape]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.target]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAnchorElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.alt]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.archive]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.code]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.codeBase]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.height]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.hspace]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.object]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.vspace]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAppletElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLAreaElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAreaElement.alt]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAreaElement.coords]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAreaElement.href]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAreaElement.shape]
TEST SUCCEEDED: The value was the empty string. [tested HTMLAreaElement.target]

TEST SUCCEEDED: The value was the empty string. [tested HTMLBaseElement.href]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBaseElement.target]

TEST SUCCEEDED: The value was the empty string. [tested HTMLBaseFontElement.color]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBaseFontElement.face]

TEST SUCCEEDED: The value was the empty string. [tested HTMLBlockquoteElement.cite]

TEST SUCCEEDED: The value was the empty string. [tested HTMLBodyElement.aLink]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBodyElement.background]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBodyElement.bgColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBodyElement.link]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBodyElement.text]
TEST SUCCEEDED: The value was the empty string. [tested HTMLBodyElement.vLink]

TEST SUCCEEDED: The value was the empty string. [tested HTMLBRElement.clear]

TEST SUCCEEDED: The value was the empty string. [tested HTMLButtonElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLButtonElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLButtonElement.value]

TEST SUCCEEDED: The value was the empty string. [tested HTMLDivElement.align]

TEST SUCCEEDED: The value was the empty string. [tested HTMLEmbedElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLEmbedElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLEmbedElement.src]
TEST SUCCEEDED: The value was the empty string. [tested HTMLEmbedElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLFontElement.color]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFontElement.face]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFontElement.size]

TEST SUCCEEDED: The value was the empty string. [tested HTMLFormElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFormElement.acceptCharset]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFormElement.action]
TEST SUCCEEDED: The value was the string 'application/x-www-form-urlencoded'. [tested HTMLFormElement.encoding]
TEST SUCCEEDED: The value was the string 'application/x-www-form-urlencoded'. [tested HTMLFormElement.enctype]
TEST SUCCEEDED: The value was the string 'get'. [tested HTMLFormElement.method]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFormElement.target]

TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.frameBorder]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.longDesc]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.marginHeight]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.marginWidth]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.scrolling]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.src]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameElement.location]

TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameSetElement.cols]
TEST SUCCEEDED: The value was the empty string. [tested HTMLFrameSetElement.rows]

TEST SUCCEEDED: The value was the empty string. [tested HTMLHeadElement.profile]

TEST SUCCEEDED: The value was the empty string. [tested HTMLHeadingElement.align]

TEST SUCCEEDED: The value was the empty string. [tested HTMLHRElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLHRElement.size]
TEST SUCCEEDED: The value was the empty string. [tested HTMLHRElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLHtmlElement.version]

TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.frameBorder]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.height]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.longDesc]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.marginHeight]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.marginWidth]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.scrolling]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.src]
TEST SUCCEEDED: The value was the empty string. [tested HTMLIFrameElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.alt]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.border]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.longDesc]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.src]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.useMap]
TEST SUCCEEDED: The value was the empty string. [tested HTMLImageElement.lowsrc]

TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.accept]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.alt]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.defaultValue]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.src]
TEST SUCCEEDED: The value was the string 'text'. [tested HTMLInputElement.type]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.useMap]
TEST SUCCEEDED: The value was the empty string. [tested HTMLInputElement.value]

TEST SUCCEEDED: The value was the empty string. [tested HTMLLabelElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLabelElement.htmlFor]

TEST SUCCEEDED: The value was the empty string. [tested HTMLLegendElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLegendElement.align]

TEST SUCCEEDED: The value was the empty string. [tested HTMLLIElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.charset]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.href]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.hreflang]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.media]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.rel]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.rev]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.target]
TEST SUCCEEDED: The value was the empty string. [tested HTMLLinkElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLMapElement.name]

TEST SUCCEEDED: The value was the empty string. [tested HTMLMetaElement.content]
TEST SUCCEEDED: The value was the empty string. [tested HTMLMetaElement.httpEquiv]
TEST SUCCEEDED: The value was the empty string. [tested HTMLMetaElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLMetaElement.scheme]

TEST SUCCEEDED: The value was the empty string. [tested HTMLModElement.cite]
TEST SUCCEEDED: The value was the empty string. [tested HTMLModElement.dateTime]

TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.code]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.archive]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.border]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.codeBase]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.codeType]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.data]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.height]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.standby]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.type]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.useMap]
TEST SUCCEEDED: The value was the empty string. [tested HTMLObjectElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLOListElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLOptGroupElement.label]

TEST SUCCEEDED: The value was the string 'null'. [tested HTMLOptionElement.text]
TEST SUCCEEDED: The value was the string 'null'. [tested HTMLOptionElement.label]
TEST SUCCEEDED: The value was the string 'null'. [tested HTMLOptionElement.value]

TEST SUCCEEDED: The value was the empty string. [tested HTMLParagraphElement.align]

TEST SUCCEEDED: The value was the empty string. [tested HTMLParamElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLParamElement.type]
TEST SUCCEEDED: The value was the empty string. [tested HTMLParamElement.value]
TEST SUCCEEDED: The value was the empty string. [tested HTMLParamElement.valueType]

TEST SUCCEEDED: The value was the empty string. [tested HTMLQuoteElement.cite]

TEST SUCCEEDED: The value was the empty string. [tested HTMLScriptElement.text]
TEST SUCCEEDED: The value was the empty string. [tested HTMLScriptElement.htmlFor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLScriptElement.event]
TEST SUCCEEDED: The value was the empty string. [tested HTMLScriptElement.charset]
TEST SUCCEEDED: The value was the empty string. [tested HTMLScriptElement.src]
TEST SUCCEEDED: The value was the empty string. [tested HTMLScriptElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLSelectElement.value]
TEST SUCCEEDED: The value was the empty string. [tested HTMLSelectElement.name]

TEST SUCCEEDED: The value was the empty string. [tested HTMLStyleElement.media]
TEST SUCCEEDED: The value was the empty string. [tested HTMLStyleElement.type]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCaptionElement.align]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.abbr]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.axis]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.bgColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.ch]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.chOff]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.headers]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.height]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.scope]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.vAlign]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableCellElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTableColElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableColElement.ch]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableColElement.chOff]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableColElement.vAlign]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableColElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.bgColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.border]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.cellPadding]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.cellSpacing]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.frame]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.rules]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.summary]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableElement.width]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTableRowElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableRowElement.bgColor]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableRowElement.ch]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableRowElement.chOff]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableRowElement.vAlign]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTableSectionElement.align]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableSectionElement.ch]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableSectionElement.chOff]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTableSectionElement.vAlign]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTextAreaElement.defaultValue]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTextAreaElement.accessKey]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTextAreaElement.name]
TEST SUCCEEDED: The value was the empty string. [tested HTMLTextAreaElement.value]

TEST SUCCEEDED: The value was the empty string. [tested HTMLTitleElement.text]

TEST SUCCEEDED: The value was the empty string. [tested HTMLUListElement.type]
